Add COS briefing and make ruff CI-green
Land COS briefing and ruff hygiene so CI matches the green local battery.

+++ packages/kernel/synelia_kernel/argent.py
@@ -19,4 +19,4 @@ def heures_vers_mois(prix_heure: int, heures: int = 730) -> int:
 
 
 def arrondi_fcfa(valeur: float) -> int:
-    return int(round(valeur))
+    return round(valeur)
